### 2018-06-27 #795 @whedon
[PRE REVIEW]: sigmajs: An R htmlwidget interface to the sigma.js visualization library
**Submitting author:** @JohnCoene (<a href="http://orcid.org/0000-0002-6637-4107">Jean-Philippe Coene</a>)
**Repository:** <a href="https://github.com/JohnCoene/sigmajs" target ="_blank">https://github.com/JohnCoene/sigmajs</a>
**Version:** v0.1.0
