### 2018-06-27 #795 @JohnCoene
Below are 5 reviewers I would suggest, based on "Preferred Programming Language" (`r`) and "Domains/topic areas you are comfortable reviewing" (`visualisation` or `networks`):

* @cdcrabtree
* @aj2duncan
* @jakobbossek
* @strengejacke
* @aj2duncan

*List from from [joss reviewers](https://docs.google.com/spreadsheets/d/1PAPRJ63yq9aPC1COLjaQp8mHmEq3rZUzwUYxTulyu78/edit#gid=856801822)*
