### 2018-06-27 #795 @strengejacke
I had a quick look at the paper, I think the topic might be suitable for me, so I'd be happy to review the submission.
