### 2018-06-28 7f265b6
export split in vignettes
