+++ vignettes/plugins.Rmd
@@ -91,11 +91,11 @@ edges <- sg_make_edges(nodes)
 sigmajs() %>% 
   sg_nodes(nodes, id, size, color) %>% 
   sg_edges(edges, id, source, target) %>% 
-  sg_export() %>% 
+  sg_export_svg() %>% 
   sg_layout() %>% 
   sg_button(
     "<i class='fa fa-download'></i>", 
-    "export", 
+    "export_svg", 
     class = "btn btn-success"
   )
 ```
@@ -110,8 +110,7 @@ edges <- sg_make_edges(nodes)
 
 sigmajs() %>% 
   sg_nodes(nodes, id, size, color) %>% 
-  sg_edges(edges, id, source, target) %>% 
-  sg_export() %>% 
+  sg_edges(edges, id, source, target) %>%
   sg_layout() %>% 
   sg_neighbours()
 ```
