### 2018-06-28 bbdfcc8
export split into svg and img functions
