+++ R/utils.R
@@ -94,7 +94,7 @@ globalVariables(c("from", "to", "."))
   c("force_start", "force_stop", "noverlap", 
     "drag_nodes", "relative_size", "add_nodes", 
     "add_edges", "drop_nodes", "drop_edges", 
-    "animate", "export")
+    "animate", "export_svg", "export_img")
 }
 
 .get_proc <- function(x){
