+++ R/buttons.R
@@ -19,7 +19,8 @@
 #'   \item{\code{drop_nodes}}
 #'   \item{\code{drop_edges}}
 #'   \item{\code{animate}}
-#'   \item{\code{export}}
+#'   \item{\code{export_svg}}
+#'   \item{\code{export_img}}
 #' }
 #' 
 #' @examples 
