+++ NEWS.md
@@ -1,3 +1,7 @@
+# sigmajs 0.1.1.9000
+
+* `sg_export` split into `sg_export_img` and `sg_export_svg`.
+
 # sigmajs 0.1.0
 
 * [crosstalk](https://rstudio.github.io/crosstalk/) integration.
