+++ NAMESPACE
@@ -30,8 +30,10 @@ export(sg_drop_nodes_delay_p)
 export(sg_drop_nodes_p)
 export(sg_edges)
 export(sg_edges2)
-export(sg_export)
-export(sg_export_p)
+export(sg_export_img)
+export(sg_export_img_p)
+export(sg_export_svg)
+export(sg_export_svg_p)
 export(sg_filter_gt_p)
 export(sg_filter_lt_p)
 export(sg_force)
