### 2018-06-28 c5aa28e
clarified CODE OF CONDUCT
