+++ README.md
@@ -56,3 +56,5 @@ sigmajs() %>%
 demo(package = "sigmajs")
 ```
 ![](pkgdown/delay.gif)
+
+Please note that this project is released with a [Contributor Code of Conduct](CODE_OF_CONDUCT.md). By participating in this project you agree to abide by its terms.
