### 2018-06-28 db0a15a
fixed crosstalk initialisation
