### 2018-06-29 39a589f
tests added
