+++ tests/testthat/test-formats.R
@@ -1,12 +1,14 @@
+library(crosstalk)
 library(testthat)
 
 context("Other formats")
 
 test_that("Gexf", {
   gexf <- system.file("examples/arctic.gexf", package = "sigmajs")
+  sd <- SharedData$new(mtcars)
   
   sg <- sigmajs() %>% 
-    sg_from_gexf(gexf)
+    sg_from_gexf(gexf, sd = sd)
   
   expect_true(sg$x$gexf)
   expect_error(sg_from_gexf())
@@ -15,11 +17,11 @@ test_that("Gexf", {
 
 test_that("igraph", {
   data("lesmis_igraph")
-  
+  sd <- SharedData$new(mtcars)
   layout <- igraph::layout_with_fr(lesmis_igraph)
   
   sg <- sigmajs() %>%
-    sg_from_igraph(lesmis_igraph, layout) %>%
+    sg_from_igraph(lesmis_igraph, layout, sd = sd) %>%
     sg_settings(defaultNodeColor = "#000")
   
   expect_length(sg$x$data, 2)
