### 2018-06-29 40e017f
example in init FUN
