+++ man/sigmajs.Rd
@@ -20,3 +20,12 @@ Initialise a graph.
 \note{
 Keep \code{width} at \code{100\%} for a responsive visualisation.
 }
+\examples{
+nodes <- sg_make_nodes()
+edges <- sg_make_edges(nodes)
+
+sigmajs() \%>\%
+  sg_nodes(nodes, id, label, size, color) \%>\%
+  sg_edges(edges, id, source, target) 
+
+}
