### 2018-06-29 dbb9e9e
select data FUN refactored
