### 2018-06-30 f11f08f
added tests
