+++ man/graph.Rd
@@ -35,7 +35,7 @@ Each edge also must include a unique \code{id} as well as two columns named \cod
 \section{Functions}{
 
 \itemize{
-	\item{Functions ending in \code{*2} take a list that resembles the original sigma.js JSON inputs}
+	\item{Functions ending in \code{2} take a list like the original sigma.js JSON.}
 	\item{Other functions take the arguments described above.}
 }
 }
