### 2018-06-30 8181956
removed useless functions
