### 2018-06-30 e0e1b17
faster dyn demo
