+++ man/camera.Rd
@@ -1,27 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/camera.R
-\name{sg_camera}
-\alias{sg_camera}
-\title{Add a camera}
-\usage{
-sg_camera(sg, elementId = NULL, initialise = FALSE)
-}
-\arguments{
-\item{sg}{An object of class \code{sigmajs}as intatiated by \code{\link{sigmajs}}.}
-
-\item{elementId}{Id of graph that initialised the camera.}
-
-\item{initialise}{Whether to initialise the camera.}
-}
-\description{
-Add a camera to your graph.
-}
-\note{
-Cameras should only be initialised once.
-}
-\examples{
-\dontrun{
-demo("add-camera", package = "sigmajs")
-}
-
-}
