### 2018-07-01 ecffe2e
renamed igraph file for consistency
