+++ man/sg_from_igraph.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/igraph.R
+% Please edit documentation in R/from_igraph.R
 \name{sg_from_igraph}
 \alias{sg_from_igraph}
 \title{Create from igraph}
