### 2018-07-01 c8a9430
input checks added
