+++ NAMESPACE
@@ -13,7 +13,6 @@ export(sg_add_nodes_delay_p)
 export(sg_add_nodes_p)
 export(sg_animate)
 export(sg_button)
-export(sg_camera)
 export(sg_clear_p)
 export(sg_cluster)
 export(sg_custom_shapes)
