### 2018-07-02 2e22d72
input tests
