+++ R/from_igraph.R
@@ -21,11 +21,10 @@
 #' @export
 sg_from_igraph <- function(sg, igraph, layout = NULL, sd = NULL) {
 
-	if (missing(sg))
-		stop("missing sg", call. = FALSE)
-
-	if (!inherits(sg, "sigmajs"))
-		stop("sg must be of class sigmajs", call. = FALSE)
+  if (missing(sg))
+    stop("missing sg", call. = FALSE)
+  
+  .test_sg(sg)
 
 	if (missing(igraph))
 		stop("must pass igraph", call. = FALSE)

[…]
