### 2018-07-02 9ccc3e7
home anim trigger earlier
