+++ R/sigmajs.R
@@ -40,7 +40,12 @@ sigmajs <- function(type = "canvas", width = "100%", height = NULL, elementId = 
     height = height,
     package = 'sigmajs',
     elementId = elementId,
-    dependencies = crosstalk::crosstalkLibs()
+    dependencies = crosstalk::crosstalkLibs(),
+    sizingPolicy = htmlwidgets::sizingPolicy(
+      defaultWidth = "100%",
+      knitr.figure = FALSE,
+      browser.fill = TRUE
+    )
   )
 }
 
