### 2018-07-03 f5c67d3
noverlap fix
