+++ man/sg_button.Rd
@@ -31,6 +31,7 @@ The default class (\code{btn btn-default}) works with Bootstrap 3, the default f
   \item{\code{relative_size}}
   \item{\code{add_nodes}}
   \item{\code{add_edges}}
+  \item{\code{add_nodes_edges}}
   \item{\code{drop_nodes}}
   \item{\code{drop_edges}}
   \item{\code{animate}}
