+++ vignettes/buttons.Rmd
@@ -58,6 +58,7 @@ Events that can be triggered via a button (corresponding function name minus `sg
 * `relative_size`
 * `add_nodes`
 * `add_edges`
+* `add_nodes_edges`
 * `drop_nodes`
 * `drop_edges`
 * `animate`
