### 2018-07-08 #795 @arfon
> I had a quick look at the paper, I think the topic might be suitable for me, so I'd be happy to review the submission.

Thanks @strengejacke. I'd like to find a second reviewer for this submission if possible so if you're happy to wait for a while before we get started I'd be much obliged :-)

@janschulz - would you be willing to review this submission for JOSS?
