### 2018-07-09 JohnCoene/sigmajs#2 @strengejacke
[JOSS-review] Add content to License file
Hi John,
this issue is related to https://github.com/openjournals/joss-reviews/issues/814.
The JOSS requires the content of a license to be included in the license file. Please add the full license text to your existing LICENSE file.

If this conflicts with a possible CRAN submission, add your license information for CRAN to the DESCRIPTION file (and probably exclude your LICENSE file from checks / package-building).
