### 2018-07-09 #814 @strengejacke
The GitHub release is already version 0.1.1, while the submitted version for review is 0.1.0 - @arfon how to deal this? Should John resubmit or edit the submission, or should we just proceed with our review?
