### 2018-07-09 #814 @arfon
> The GitHub release is already version 0.1.1, while the submitted version for review is 0.1.0 - @arfon how to deal this? Should John resubmit or edit the submission, or should we just proceed with our review?

If it's OK with @JohnCoene, it probably makes sense to review the latest code and just update the JOSS submission (I can do that here).
