### 2018-07-09 JohnCoene/sigmajs#3 @strengejacke
[JOSS-review] statement of need
Related to https://github.com/openjournals/joss-reviews/issues/814

The README file contains a paragraph on the "statement of need", which describes very well who may benefit from this package.

Please revise the introduction of your paper and add a similar statement of need to it, as the summary paragraph does not fully satisfy the requirements for the statement of need. This is just a minor issue, you may copy from your readme.

One addition to the readme-file. You write

> sigmajs leverages the original library's great many methods

This is unclear to user who are not familiar with sigma.js, as it's not clear what the "original libraries" are. Please add a small sentence, which explains what the original library does and why there's an R implementation. **Edit:** _There's a good example in your paper. Maybe you should add some parts from the paper to the readme and vice versa._
