### 2018-07-09 JohnCoene/sigmajs#4 @strengejacke
[JOSS-review] Guidelines for how to contribute
related to https://github.com/openjournals/joss-reviews/issues/814

The documentation requires guidelines for community members to
1) Contribute to the software
2) Report issues or problems with the software
3) Seek support

Please add this information to the README, or if these instructions will be too long for the readme and obscure the introduction into the package, please add a (markdown) file and link from the readme to that file.
