### 2018-07-09 JohnCoene/sigmajs#5 @strengejacke
[JOSS-review] Broken link to Code of Conduct
related to https://github.com/openjournals/joss-reviews/issues/814

The link to the code-of-conduct in the readme is broken and needs to be fixed.
