### 2018-07-10 #814 @strengejacke
I hope to finish the last parts of my review today, as I'm on holiday tomorrow and be back after 29th of July.
