+++ README.md
@@ -57,4 +57,4 @@ demo(package = "sigmajs")
 ```
 ![](pkgdown/delay.gif)
 
-Please note that this project is released with a [Contributor Code of Conduct](CODE_OF_CONDUCT.md). By participating in this project you agree to abide by its terms.
+Please note that this project is released with a [Contributor Code of Conduct](CONDUCT.md). By participating in this project you agree to abide by its terms.
