### 2018-07-12 JohnCoene/sigmajs#2 @JohnCoene
I have used the [License content from the original sigma.js library](https://github.com/jacomyal/sigma.js/blob/master/LICENSE.txt) (MIT) as it should follow the same licensing.
