+++ CONTRIBUTING.md
@@ -1,5 +1,9 @@
 # Contributing to sigmajs
 
+## Support
+
+If you have a question on how to use the package please first explore the [online documentation website](http://sigmajs.john-coene.com/), you may also want to check if your question has already been asked and answered on [stackoverflow](https://stackoverflow.com/) before filing an issue. In the latter case please read the section below before doing so.
+
 ## Filing an issue
 
 You are free to open an [issue](https://github.com/JohnCoene/sigmajs/issues) if you encounter a problem with the package, where possible, illustrate the said problem with a [reproducible example](https://www.tidyverse.org/help/#reprex).
