+++ JOSS/paper.bib
@@ -72,3 +72,11 @@
   note = {R package version 1.1.0},
   url = {https://CRAN.R-project.org/package=pkgdown},
 }
+
+@Manual{,
+  title = {magrittr: A Forward-Pipe Operator for R},
+  author = {Stefan Milton Bache and Hadley Wickham},
+  year = {2014},
+  note = {R package version 1.5},
+  url = {https://CRAN.R-project.org/package=magrittr},
+}
