+++ README.md
@@ -57,4 +57,6 @@ demo(package = "sigmajs")
 ```
 ![](pkgdown/delay.gif)
 
-Please note that this project is released with a [Contributor Code of Conduct](CONDUCT.md). By participating in this project you agree to abide by its terms.
+## Contributing
+
+See the [contrinuting guidelines](CONTRIBUTING.md) if you encounter any issue. Please note that this project is released with a [Contributor Code of Conduct](CONDUCT.md). By participating in this project you agree to abide by its terms.
