### 2018-07-13 JohnCoene/sigmajs#3 @JohnCoene
I think I have remedied to the above but somewhat unsure if I did it properly.
