### 2018-07-15 JohnCoene/sigmajs#3 @jankatins
I second this.

Coming from igraph, I would have also liked to see what the benefit of this library is over igraph (interactivity?)
