### 2018-07-15 JohnCoene/sigmajs#6 @jankatins
Does not display in jupyter notebook nor in rstudio
I currently can't get output in neither jupyter notebook (In irkernel I get `ERROR while rich displaying an object: Error in file(file, "rb"): invalid 'description' argument`) nor in Rstudio (latest download -> as far as I can identifiy it it's a empty html plot displayed in the plot area of rstudio). I followed the install from github instruction and using the latest R from debian unstable.

I tried this code:
```r
library(sigmajs)

# generate data
nodes <- sg_make_nodes()
edges <- sg_make_edges(nodes)

# visualise
sigmajs() %>%
  sg_nodes(nodes, id, label, size, color) %>%
  sg_edges(edges, id, source, target)
```
I'm not sure what is happening here, is there anything I can do to debug this?

Here is the complete error in IRkernel:

```R
ERROR while rich displaying an object: Error in file(file, "rb"): invalid 'description' argument

Traceback:
1. FUN(X[[i]], ...)
2. tryCatch(withCallingHandlers({
 .     rpr <- mime2repr[[mime]](obj)
 .     if (is.null(rpr))
 .         return(NULL)
 .     prepare_content(is.raw(rpr), rpr)
 . }, error = error_handler), error = outer_handler)
3. tryCatchList(expr, classes, parentenv, handlers)
4. tryCatchOne(expr, names, parentenv, handlers[[1L]])
5. doTryCatch(return(expr), name, parentenv, handler)
6. withCallingHandlers({
 .     rpr <- mime2repr[[mime]](obj)
 .     if (is.null(rpr))
 .         return(NULL)
 .     prepare_content(is.raw(rpr), rpr)
 . }, error = error_handler)
7. mime2repr[[mime]](obj)
8. repr_html.htmlwidget(obj)
9. lapply(obj$dependencies, function(dep) {
 .     html <- c()
 .     if (length(dep$script) > 0) {
 .         f <- file.path(dep$src$file, dep$script)
 .         html <- c(html, sprintf("<script src=\"%s\"></script>",
 .             dataURI(mime = "application/javascript", file = f)))
 .     }
 .     if (length(dep$stylesheet) > 0) {
 .         f <- file.path(dep$src$file, dep$stylesheet)
 .         html <- c(html, sprintf("<link href=\"%s\" rel=\"stylesheet\" />",
 .             dataURI(mime = "text/css;charset-utf-8", file = f)))
 .     }
 .     paste(html, collapse = "\n")
 . })
10. FUN(X[[i]], ...)
11. sprintf("<script src=\"%s\"></script>", dataURI(mime = "application/javascript",
  .     file = f))
12. dataURI(mime = "application/javascript", file = f)
13. file(file, "rb")
```

The Rstudio output looks like this:

![image](https://user-images.githubusercontent.com/890156/42736622-fa037e12-8868-11e8-846b-573690a3fd0e.png)
