### 2018-07-15 #814 @jankatins
I currently can't get output in neither jupyter notebook nor in Rstudio. I opened https://github.com/JohnCoene/sigmajs/issues/6 to track the issue.
