### 2018-07-15 JohnCoene/sigmajs#6 @JohnCoene
There is no output in Rstudio viewer, there is a note on that in the README. Open in browser to view. It's an issue with Rstudio, not the widget.
