### 2018-07-15 JohnCoene/sigmajs#6 @JohnCoene
Should I add a note in every function? I'm unsure as to how to make this clear to the user.
