### 2018-07-18 JohnCoene/sigmajs#6 @jankatins
Can someone elaborate what the issue is with that? Usually htmlwidgets work fine in RStudio...

IMO, a package which should plot but does not show a plot in RStudio is kind of broken (I can accept that it doesn't work in jupyter R kernels, the ecosystem around that is unfortunately not on the same level :-( ).
