### 2018-07-19 JohnCoene/sigmajs#6 @JohnCoene
The plots do not show because the package does not use the standard htmlwidget DOM, in order to add buttons. The plots do work
