### 2018-07-23 d07c170
checks for RStudio input + open in browser by default
