### 2018-07-23 JohnCoene/sigmajs#6 @JohnCoene
On this issue, I have just pushed changes with I think somewhat remedy to this, please tell me what you think.

1. Visualisation automatically open in external browser.
2. When using RStudio a `warning` is sent to the console stating the above (opening in external browser)
