+++ index.Rmd
@@ -243,7 +243,7 @@ devtools::install_bitbucket("JohnCoene/sigmajs") # bitbucket
 
 Most functions have corresponding `demo()`.
 
-*Note that the graphs may not work in RStudio viewer, open them in the browser of your choice.*
+*Note that the graphs do not work in the RStudio viewer, and thus open in your default browser.*
 
 ```{r, eval=TRUE}
 library(sigmajs)
