### 2018-07-25 ed15d9d
minor improvements
