+++ JOSS/paper.md
@@ -18,16 +18,16 @@ affiliations:
 
 With the rise in popularity of networks, it is important for R users to have access to a package that allows visualising the aforementioned networks in a highly configurable, interactive and dynamic manner. `sigmajs` is a fully-fledged wrapper for the [sigma.js JavaScript library](http://sigmajs.org/).
 
-The sigma.js JavaScript library is described as follows on its [website]((http://sigmajs.org/)): 
+The sigma.js JavaScript library is described as follows on its [website](http://sigmajs.org/): 
 
 > Sigma is a JavaScript library dedicated to graph drawing. It makes easy to publish networks on Web pages, and allows developers to integrate network exploration in rich Web applications.
 
 The package `sigmajs` [@sigmajs] bridges the sigma.js JavaScript 
 library and R [@R2018] via the `htmlwidgets` package [@htmlwidgets]. The package also 
 extends the original JavaScript library by providing additional functions, namely using 
-the `igraph` package [@igraph] that enable the user to layout and cluster graphs. 
+the `igraph` package [@igraph] to enable the user to layout and cluster graphs. 
 Finally `sigmajs` is also integrated with the `crosstalk` package [@crosstalk] which 
-enables graphs to be wired to other `htmlwidgets` such as `plotly` [@plotly] and 
+lets graphs be wired to other `htmlwidgets` such as `plotly` [@plotly] and 
 `leafelt` [@leaflet].
 
 # Functionality
@@ -36,7 +36,7 @@ Graphs are initialised with `sigmajs()`, all other functions start with `sg_`, a
 
 * Shiny proxies.
 * Crosstalk integration.
-* Buttons to trigger various animations.
+* Buttons to trigger various animations and events.
 * Possibility to capture graph interactions in Shiny.
 
 The R package `sigmajs` is available on [GitHub](https://github.com/JohnCoene/sigmajs) or 
