+++ DESCRIPTION
@@ -2,7 +2,7 @@ Package: sigmajs
 Title: Interface to 'Sigma.js' Graph Visualization Library
 Date: 2018-06-26
 Version: 0.1.1.9000
-Authors@R: person("John", "Coene", email = "jcoenep@gmail.com", role = c("aut", "cre"))
+Authors@R: person("John", "Coene", email = "jcoenep@gmail.com", role = c("aut", "cre"), comment = c(ORCID = "0000-0002-6637-4107"))
 Description: Interface to 'sigma.js' graph visualization library including animations, plugins and shiny proxies.
 License: MIT + file LICENSE
 Encoding: UTF-8
