### 2018-07-31 d4d2cca
multiple events trigger on button
