### 2018-08-02 JohnCoene/sigmajs#3 @strengejacke
> I think I have remedied to the above but somewhat unsure if I did it properly.

I would also add a stronger statement of need to the README file - you can copy from your _paper.md_. Then I personally would close this issue.
