### 2018-08-02 JohnCoene/sigmajs#6 @strengejacke
Works for me.
