### 2018-08-02 #814 @strengejacke
From my perspective, just a minor issue left (statement of need in readme-file, see https://github.com/JohnCoene/sigmajs/issues/3).
