### 2018-08-02 2f14d7f
correct reordering in cluster and layout
