### 2018-08-02 8e52700
updated NEWS
