+++ NEWS.md
@@ -6,6 +6,7 @@
 * `sg_get_layout` returns nodes with coordinates rather than just the coordinates.
 * Graphs open in browser by default + warning is sent to console when using RStudio.
 * Added ability to trigger multiple events from the button.
+* Fixed bug in cluster and layout function where igraph would error on duplicate vertex name.
 
 # sigmajs 0.1.0
 
