### 2018-08-02 #814 @JohnCoene
Thank you @strengejacke, I have added the statement of need from the paper to the README: https://github.com/JohnCoene/sigmajs/commit/550e5809da3973100f3d10a80edb6f85866445c2
