### 2018-08-06 9f31478
logo size
