### 2018-08-06 599bcaa
logo height fixed
