### 2018-08-06 180ed01
homepage redone + get started link to layout
