### 2018-08-06 fd3f8dd
corrections on homepage
