+++ R/helpers.R
@@ -5,6 +5,15 @@
 #' @param n Number of nodes.
 #' @param colors Color palette to use.
 #' @param nodes Nodes, as generated by \code{sg_make_nodes}.
+#' @param k the degree of each vertex in the graph.
+#' @param ... Any other argument to pass to \link[igraph]{sample_pa}.
+#' 
+#' @section Functions:
+#' \itemize{
+#'   \item{\code{sg_make_nodes} generate data.frame nodes.}
+#'   \item{\code{sg_make_edges} generate data.frame edges.}
+#'   \item{\code{sg_make_nodes_edges} generate list of nodes and edges.}
+#' }
 #'
 #' @examples
 #' nodes <- sg_make_nodes()
@@ -15,7 +24,7 @@
 #'	sg_edges(edges, id, source, target) %>%
 #'	sg_settings(defaultNodeColor = "#0011ff")
 #' 
-#' @return \code{tibble} of nodes or edges.
+#' @return \code{tibble} of nodes or edges or a \code{list} of the latter.
 #'
 #' @importFrom grDevices colorRampPalette
 #' @rdname generate
@@ -44,4 +53,44 @@ sg_make_edges <- function(nodes, n = nrow(nodes) * 1.5) {
 		source = sample(ids, n, replace = TRUE),
 		target = sample(ids, n, replace = TRUE)
 	)
+}
+
+#' @rdname generate
+#' @export
+sg_make_nodes_edges <- function(n, ...){
+  
+  g <- igraph::sample_pa(n, ...)
+  
+  g <- igraph::as_data_frame(g, what = 'both')
+  
+  edges <- g$edges %>% 
+    dplyr::mutate(
+      id = 1:n(),
+      from = as.character(from),
+      to = as.character(to)
+    ) %>%
+    dplyr::select(
+      source = from, 
+      target = to,
+      id)
+  
+  nodes <- dplyr::tibble(
+    id = as.character(c(edges$source, edges$target))
+  ) %>% 
+    dplyr::group_by(id) %>% 
+    dplyr::summarise(size = n()) %>% 
+    dplyr::ungroup() %>% 
+    dplyr::mutate(
+      label = id
+    ) %>% 
+    dplyr::select(
+      id, 
+      label, 
+      size
+    )
+  
+  list(
+    nodes = nodes,
+    edges = edges
+  )
 }
