+++ man/generate.Rd
@@ -3,12 +3,15 @@
 \name{sg_make_nodes}
 \alias{sg_make_nodes}
 \alias{sg_make_edges}
+\alias{sg_make_nodes_edges}
 \title{Generate data}
 \usage{
 sg_make_nodes(n = 10, colors = c("#B1E2A3", "#98D3A5", "#328983", "#1C5C70",
   "#24C96B"))
 
 sg_make_edges(nodes, n = nrow(nodes) * 1.5)
+
+sg_make_nodes_edges(n, k, directed = FALSE, multiple = FALSE)
 }
 \arguments{
 \item{n}{Number of nodes.}
@@ -16,6 +19,12 @@ sg_make_edges(nodes, n = nrow(nodes) * 1.5)
 \item{colors}{Color palette to use.}
 
 \item{nodes}{Nodes, as generated by \code{sg_make_nodes}.}
+
+\item{k}{the degree of each vertex in the graph.}
+
+\item{directed}{Whether graph should be directed.}
+
+\item{multiple}{Whether multiple edges are allowed.}
 }
 \value{
 \code{tibble} of nodes or edges.
@@ -23,6 +32,15 @@ sg_make_edges(nodes, n = nrow(nodes) * 1.5)
 \description{
 Generate nodes and edges.
 }
+\section{Functions}{
+
+\itemize{
+  \item{\code{sg_make_nodes} generate data.frame nodes.}
+  \item{\code{sg_make_edges} generate data.frame edges.}
+  \item{\code{sg_make_nodes_edges} generate list of nodes and edges.}
+}
+}
+
 \examples{
 nodes <- sg_make_nodes()
 edges <- sg_make_edges(nodes)
