+++ NAMESPACE
@@ -51,6 +51,7 @@ export(sg_kill_p)
 export(sg_layout)
 export(sg_make_edges)
 export(sg_make_nodes)
+export(sg_make_nodes_edges)
 export(sg_neighbors)
 export(sg_neighbours)
 export(sg_nodes)
