### 2018-08-06 JohnCoene/sigmajs#6 @jankatins
Works for me -> closing
