### 2018-08-06 #814 @jankatins
From my perspective, all points are checked in master (the statement of need is added to the readme). I assume the version requirement in this issue can be fixed somehow, but it needs a new release of sigmajs first?

> Functions can be piped, referring to the magrittr package [@magrittr] pipe operator (%>%), to build the desired graph.

This sounds a bit strange (to my german-english ears, so take it with bit of care): "Functions can be piped (with `%>%` from the  magrittr package [@magrittr]) to build the desired graph" or "Functions can be piped with the standard tidyverse pipe operator `%>%` [@magrittr] to build the desired graph"
