### 2018-08-06 #814 @strengejacke
All my raised issues are addressed adequately, too. I recommend to accept the paper.
