### 2018-08-07 #814 @arfon
@JohnCoene - do you want to make a final release based on the changes from this review?
