+++ man/generate.Rd
@@ -11,7 +11,7 @@ sg_make_nodes(n = 10, colors = c("#B1E2A3", "#98D3A5", "#328983", "#1C5C70",
 
 sg_make_edges(nodes, n = nrow(nodes) * 1.5)
 
-sg_make_nodes_edges(n, k, directed = FALSE, multiple = FALSE)
+sg_make_nodes_edges(n, ...)
 }
 \arguments{
 \item{n}{Number of nodes.}
@@ -20,14 +20,12 @@ sg_make_nodes_edges(n, k, directed = FALSE, multiple = FALSE)
 
 \item{nodes}{Nodes, as generated by \code{sg_make_nodes}.}
 
+\item{...}{Any other argument to pass to \link[igraph]{sample_pa}.}
+
 \item{k}{the degree of each vertex in the graph.}
-
-\item{directed}{Whether graph should be directed.}
-
-\item{multiple}{Whether multiple edges are allowed.}
 }
 \value{
-\code{tibble} of nodes or edges.
+\code{tibble} of nodes or edges or a \code{list} of the latter.
 }
 \description{
 Generate nodes and edges.
