+++ NAMESPACE
@@ -38,6 +38,7 @@ export(sg_filter_lt_p)
 export(sg_force)
 export(sg_force_config_p)
 export(sg_force_kill_p)
+export(sg_force_restart)
 export(sg_force_restart_p)
 export(sg_force_start)
 export(sg_force_start_p)

[…]
