### 2018-08-07 6772cdc
added missing docs
