+++ R/helpers.R
@@ -5,7 +5,6 @@
 #' @param n Number of nodes.
 #' @param colors Color palette to use.
 #' @param nodes Nodes, as generated by \code{sg_make_nodes}.
-#' @param k the degree of each vertex in the graph.
 #' @param ... Any other argument to pass to \link[igraph]{sample_pa}.
 #' 
 #' @section Functions:
