+++ man/generate.Rd
@@ -21,8 +21,6 @@ sg_make_nodes_edges(n, ...)
 \item{nodes}{Nodes, as generated by \code{sg_make_nodes}.}
 
 \item{...}{Any other argument to pass to \link[igraph]{sample_pa}.}
-
-\item{k}{the degree of each vertex in the graph.}
 }
 \value{
 \code{tibble} of nodes or edges or a \code{list} of the latter.
