### 2018-08-07 2eb9b85
add edges delay JS corrected
