### 2018-08-07 8c6d912
drop edge working
