### 2018-08-08 c313d10
sg_force_restart fixed (at last)
