### 2018-08-08 7aa8688
updated version ahead of CRAN JOSS paper
