+++ pkgdown/extra.js
@@ -12,9 +12,3 @@ $('#dynamism a').click(function (e) {
   e.preventDefault()
   $(this).tab('show')
 })
-
-var alerted = localStorage.getItem('alerted') || '';
-if (alerted != 'yes') {
- alert("My alert.");
- localStorage.setItem('alerted','yes');
-}
