### 2018-08-08 d5557ac
tests match sg_neightbours new args
