### 2018-08-08 84d0336
crosstalk intergration JS improved
