+++ inst/htmlwidgets/sigmajs.js
@@ -373,126 +373,6 @@ HTMLWidgets.widget({
 				  }
 				}
 
-				// events
-				if (HTMLWidgets.shinyMode) {
-					// click node
-					s.bind('clickNode', function (e) {
-						Shiny.onInputChange(el.id + '_click_node' + ":sigmajsParse", e.data.node);
-					});
-
-					// click nodeS
-					s.bind('clickNodes', function (e) {
-						Shiny.onInputChange(el.id + '_click_nodes' + ":sigmajsParse", e.data.nodes);
-					});
-
-					// click edge
-					s.bind('clickEdge', function (e) {
-						Shiny.onInputChange(el.id + '_click_edge' + ":sigmajsParse", e.data.edge);
-					});
-
-					// click edgeS
-					s.bind('clickEdges', function (e) {
-						Shiny.onInputChange(el.id + '_click_edges' + ":sigmajsParse", e.data.edges);
-					});
-
-					// click stage
-					s.bind('clickStage', function (e) {
-						Shiny.onInputChange(el.id + '_click_stage' + ":sigmajsParse", true);
-					});
-
-					// double click stage
-					s.bind('doubleClickStage', function (e) {
-						Shiny.onInputChange(el.id + '_double_click_stage' + ":sigmajsParse", true);
-					});
-
-					// right click stage
-					s.bind('rightClickStage', function (e) {
-						Shiny.onInputChange(el.id + '_right_click_stage' + ":sigmajsParse", true);
-					});
-
-					// double click node
-					s.bind('doubleClickNode', function (e) {
-						Shiny.onInputChange(el.id + '_double_click_node' + ":sigmajsParse", e.data.node);
-					});
-
-					// double click nodeS
-					s.bind('doubleClickNodes', function (e) {
-						Shiny.onInputChange(el.id + '_double_click_nodes' + ":sigmajsParse", e.data.nodes);
-					});
-
-					// double click edge
-					s.bind('doubleClickEdge', function (e) {
-						Shiny.onInputChange(el.id + '_double_click_edge' + ":sigmajsParse", e.data.edge);
-					});
-
-					// double click edgeS
-					s.bind('doubleClickEdges', function (e) {
-						Shiny.onInputChange(el.id + '_double_click_edges' + ":sigmajsParse", e.data.edges);
-					});
-
-					// right click node
-					s.bind('rightClickNode', function (e) {
-						Shiny.onInputChange(el.id + '_right_click_node' + ":sigmajsParse", e.data.node);
-					});
-
-					// right click nodeS
-					s.bind('rightClickNodes', function (e) {
-						Shiny.onInputChange(el.id + '_right_click_nodes' + ":sigmajsParse", e.data.nodes);
-					});
-
-					// right click edge
-					s.bind('rightClickEdge', function (e) {
-						Shiny.onInputChange(el.id + '_right_click_edge' + ":sigmajsParse", e.data.edge);
-					});
-
-					// right click edgeS
-					s.bind('rightClickEdges', function (e) {
-						Shiny.onInputChange(el.id + '_right_click_edges' + ":sigmajsParse", e.data.edges);
-					});
-
-					// over node
-					s.bind('overNode', function (e) {
-						Shiny.onInputChange(el.id + '_over_node' + ":sigmajsParse", e.data.node);
-					});
-
-					// over nodeS
-					s.bind('overNodes', function (e) {
-						Shiny.onInputChange(el.id + '_over_nodes' + ":sigmajsParse", e.data.nodes);
-					});
-
-					// over edge
-					s.bind('overEdge', function (e) {
[…]
