+++ R/zzz.R
@@ -1,11 +1,11 @@
 .onAttach <- function(libname, pkgname) {
-	shiny::registerInputHandler("sigmajsParse", function(data, ...) {
+	shiny::registerInputHandler("sigmajsParseJS", function(data, ...) {
 		jsonlite::fromJSON(jsonlite::toJSON(data, auto_unbox = TRUE))
 	}, force = TRUE)
 }
 
 .onLoad <- function(libname, pkgname) {
-	shiny::registerInputHandler("sigmajsParse", function(data, ...) {
+	shiny::registerInputHandler("sigmajsParseJS", function(data, ...) {
 		jsonlite::fromJSON(jsonlite::toJSON(data, auto_unbox = TRUE))
 	}, force = TRUE)
 }
