### 2018-08-09 8091629
added kill argument for shiny
