### 2018-08-09 #814 @JohnCoene
@arfon Updated version https://github.com/JohnCoene/sigmajs/commit/7aa86886df65e58a82d7cb983c3db29e10fd1565, I believe the dev (`.9000`) did not meet the requirements: sigmajs is now on `0.1.1`.
