### 2018-08-10 #814 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
